+++ ui-wear-compose-material3/src/main/java/com/github/droibit/oss_licenses/ui/wear/compose/material3/internal/OssLicenseList.kt
@@ -1,13 +1,8 @@
 package com.github.droibit.oss_licenses.ui.wear.compose.material3.internal
 
 import androidx.annotation.RestrictTo
-import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.platform.LocalConfiguration
-import androidx.compose.ui.platform.LocalDensity
-import androidx.compose.ui.unit.Dp
-import androidx.compose.ui.unit.dp
 import androidx.wear.compose.foundation.ExperimentalWearFoundationApi
 import androidx.wear.compose.foundation.lazy.TransformingLazyColumn
 import androidx.wear.compose.foundation.lazy.TransformingLazyColumnItemScope
@@ -15,7 +10,7 @@ import androidx.wear.compose.foundation.lazy.TransformingLazyColumnState
 import androidx.wear.compose.foundation.lazy.items
 import androidx.wear.compose.foundation.lazy.rememberTransformingLazyColumnState
 import com.github.droibit.oss_licenses.parser.OssLicense
-import kotlin.math.ceil
+import com.github.droibit.oss_licenses.ui.wear.compose.core.rememberResponsiveColumnPadding
 
 @RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
 @OptIn(ExperimentalWearFoundationApi::class)
@@ -42,30 +37,5 @@ fun OssLicenseList(
       key = OssLicense::library,
       itemContent = listItem,
     )
-  }
-}
-
-@Composable
-fun rememberResponsiveColumnPadding(): PaddingValues {
-  val configuration = LocalConfiguration.current
-  val screenHeightDp = configuration.screenHeightDp.dp
-  val screenWidthDp = configuration.screenWidthDp.dp
-
-  val horizontalPadding = (screenWidthDp * 0.052f).ceilPx()
-
-  return PaddingValues(
-    top = (screenHeightDp * 0.1664f).ceilPx(),
-    bottom = (screenHeightDp * 0.3646f).ceilPx(),
-    start = horizontalPadding,
-    end = horizontalPadding,
-  )
-}
-
-@Composable
-internal fun Dp.ceilPx(): Dp {
-  val density = LocalDensity.current
-
-  return with(density) {
-    ceil(this@ceilPx.toPx()).toDp()
   }
 }
